Fix error that caused some FN to not be counted
Some video LabelRows don't have frames with default values
on the case that no classification nor object was detected
in a specific frame, so it's possible that baseline_project
has no info in a frame and other projects do. That case wasn't
checked and some false negatives were not counted.