DEV-2084 - Allow adding flat classifications to COCO exported file (#217)
One test has been added.

Support for the `dynamic` field in the ontology has been added. This needs to be added to the `master` branch too.

All `DENIS:` comments have been renamed as `TODO:` comments.

Essentially all non-nested attribute answers are added within each label within `attributes` within the `classifications`. 
This means that we're copying a bit of data - however it is more in line with what CVAT does. I am always adding defaults with `None` answers if appropriate. 
So if a user never toggles anything then
* the text answer shows up as `None`
* the radio answer shows up as `None`
* the checklist answer shows up as `False`

otherwise 
* the text answers shows up as a string
* the radio answer shows up as a string the same as name of the option chosen
* the checklsit answer shows up as `True` or `False`

Name clashes are somewhat taken care of, but I do not expect this to happen for most cases so the 80/20 thing for me was to ensure that it doesn't crash if that happens and let clients ask for this extension in which case I'll add a feature_hash somewhere when there is a name clash.