Add method to verify a service integration token to user client
This commit adds a new `ServiceIntegration` object to the SDK, that
interacts with service integrations in Annotate via the API.

Initially, only a `verify_token` method has been added - which when
called either returns a verified JWT, or raises an error.

The `EncordUserClient` has been given a new `get_service_integration`
method, that returns the new `ServiceIntegration` object. Note that this
does not currently make an API request, because it does not need to.