feat: Allow to have variable trace_id_provider and stateful.
If we want to use the Encord SDK client for a multi-request transaction,
then in order to tag logs together appropriately, then we need a method of recording
the current span id and updating it in response to what the server says.
If the server tells us we are on Span 7, we should update appropriately.

Bit of weirdness going on here.
Having the trace_id_provider historically also include the span_id means that this is best handled via some closure that mutates some state that seeds the trace_id_provider.

This way gives an interface that allows the code to call the span id store and update appropriately.
This means that we will only be incrementing the span_id if explicitly told too. So we should support all cases.
Very hairy